+++ kiva-frontend/next.config.ts
